Add shared memory transport layer for the stm32wb55 to use connection and services the BLE RF through IPM SharedMem
The Rx data are stored in a fifo to manage connection, and ACL packets
with context section to protect index when receiving data.
The BLE init set the device address and power settings just after the HCI_RESET command sent
The LSE clock is used for the RF and WakeUp capability by the variant.

Signed-off-by: Francois Ramu <francois.ramu@st.com>